creates new replay_buffers directory
